requirements.txt
tensorflow>=2.15.0
streamlit>=1.32.0
numpy>=1.24.0
Pillow>=9.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
kaggle>=1.5.0
